% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy}
\usage{
accuracy(expected, observed)
}
\arguments{
\item{expected}{a vector of logical values for the expected values}

\item{observed}{a vector of logical values for the observed values}
}
\value{
a list of accuracy parameters
}
\description{
Signal detection values for modules that classify papers as having a feature or not
}
