% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{collapse_section}
\alias{collapse_section}
\title{Make Collapsible Section}
\usage{
collapse_section(text, title = "Learn More")
}
\arguments{
\item{text}{The text to put in the collapsible section; vectors will be collapse with line breaks between (e.g., into paragraphs)}

\item{title}{The title of the collapse header}
}
\value{
text
}
\description{
A helper function for making module reports.
}
\examples{
text <- c("Paragraph 1...", "Paragraph 2...")
collapse_section(text) |> cat()
}
