% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_predict.R
\name{distinctive_words}
\alias{distinctive_words}
\title{Get Distinctive Words}
\usage{
distinctive_words(text, classification, n = 15)
}
\arguments{
\item{text}{a vector of text to assess}

\item{classification}{a vector of the classification}

\item{n}{the number of top distnctive words to get}
}
\value{
a data frame of the words
}
\description{
Get Distinctive Words
}
