% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.R
\name{doi_clean}
\alias{doi_clean}
\title{Clean DOIs}
\usage{
doi_clean(doi)
}
\arguments{
\item{doi}{a character vector of one or more DOIs}
}
\value{
a character vector of cleaned DOIs (no https://doi.org or DOI:)
}
\description{
Clean DOIs
}
\examples{
doi_clean("https://doi.org/10.1038/nphys1170")
doi_clean("doi:10.1038/nphys1170")
doi_clean("DOI: 10.1038/nphys1170")
}
