% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.R
\name{doi_resolves}
\alias{doi_resolves}
\title{Check whether a DOI resolves}
\usage{
doi_resolves(doi, timeout = 10)
}
\arguments{
\item{doi}{Character vector. One or more DOIs to check.}

\item{timeout}{Numeric. Request timeout in seconds. Default is \code{10}.}
}
\value{
Logical vector. For each input DOI, returns TRUE if the DOI resolves,
FALSE if it does not resolve (or does not start with 10.), and NA if the check failed.
}
\description{
Checks the doi.org API to see if a DOI is registered and has an associated URL
(using \verb{https://doi.org/api/handles}). Returns TRUE if it does, FALSE if the DOI
does not exist or does not have an associated URL, and NA if the test failed.
Clearly invalid DOIs (i.e. not starting with "10.") will return FALSE without
server requests.
}
\examples{
\dontrun{
doi_resolves("10.1038/nphys1170") # Expected: TRUE
doi_resolves("10.1234/invalid.doi") # Expected: FALSE
}
}
