% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.R
\name{doi_valid_format}
\alias{doi_valid_format}
\title{Validate DOI format}
\usage{
doi_valid_format(doi)
}
\arguments{
\item{doi}{a character vector of one or more DOIs}
}
\value{
a logical vector
}
\description{
Validate DOI format
}
\examples{
doi_valid_format("10.1038/nphys1170")
doi_valid_format("no.no.10.1038")
}
