% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication_database.R
\name{FReD}
\alias{FReD}
\title{FORRT Replication Database}
\format{
A data frame with 2222 rows and 4 columns:
\describe{
\item{ref_original}{reference of original study}
\item{doi_original}{doi of original study}
\item{ref_replication}{reference of replication study}
\item{ref_replication}{doi of replication study}
}
}
\source{
\url{https://osf.io/9r62x/files/z5u9b}
}
\usage{
FReD()
}
\value{
a data frame
}
\description{
FReD database containing DOIs of original studies and replications. Use \code{fred_date()} to find the date it was downloaded, and \code{fred_update()} to update it.
}
\examples{
FReD()
}
