% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication_database.R
\name{FReD_update}
\alias{FReD_update}
\title{Update FReD}
\usage{
FReD_update()
}
\value{
the path to the data frame (invisibly)
}
\description{
metacheck comes with a built-in data frame called \code{FReD}. We update it regularly, but you can use this function to download the newest version. The download is >5MB, but this function will summarise the information into a smaller version and delete the original file.
}
