% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref.R
\name{get_doi}
\alias{get_doi}
\title{Get DOI from Reference}
\usage{
get_doi(reference, min_score = 50)
}
\arguments{
\item{reference}{the full text reference of the paper to get info for}

\item{min_score}{minimal score that is taken to be a reliable match (default 50)}
}
\value{
doi
}
\description{
Get DOI from Reference
}
\examples{
ref <- paste(
  "Lakens, D., Mesquida, C., Rasti, S., & Ditroilo, M. (2024).",
  "The benefits of preregistration and Registered Reports.",
  "Evidence-Based Toxicology, 2(1)."
)
\donttest{
  doi <- get_doi(ref)
}
}
