% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacheck.R
\docType{package}
\name{metacheck-package}
\alias{metacheck}
\alias{metacheck-package}
\title{metacheck: Check Research Outputs for Best Practices}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A modular, extendable system for automatically checking research outputs for best practices using text search, R code, and/or (optional) LLM queries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/scienceverse/metacheck}
  \item \url{https://scienceverse.github.io/metacheck/}
  \item Report bugs at \url{https://github.com/scienceverse/metacheck/issues}
}

}
\author{
\strong{Maintainer}: Lisa DeBruine \email{debruine@gmail.com} (\href{https://orcid.org/0000-0002-7523-5539}{ORCID})

Authors:
\itemize{
  \item Daniel Lakens \email{lakens@gmail.com} (\href{https://orcid.org/0000-0002-0247-239X}{ORCID})
}

Other contributors:
\itemize{
  \item Cristian Mesquida \email{cristianmesquida@gmail.com} (\href{https://orcid.org/0000-0002-1542-8355}{ORCID}) [contributor]
  \item Jakub Werner \email{jakub@jakubwerner.com} (\href{https://orcid.org/0009-0004-0388-0362}{ORCID}) [contributor]
  \item Lukas Wallrich \email{l.wallrich@bbk.ac.uk} (\href{https://orcid.org/0000-0003-2121-5177}{ORCID}) [contributor]
}

}
\keyword{internal}
