% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{op-vect-null-default}
\alias{op-vect-null-default}
\alias{rep_if}
\title{Replace If}
\usage{
rep_if(x, y, replace = NULL)
}
\arguments{
\item{x}{For each \code{x[[i]]} if NULL, return \code{y[[i]]}; otherwise return \code{x[[i]]}.}

\item{y}{Replacement value(s); if not the same length as \code{x}, then values will be recycled}
}
\description{
Replace values if NULL, NA, or sepcified value
}
\examples{
list(NULL, 1) \%|||\% 2
list(NULL, 0, NULL) \%|||\% 1:3
}
\keyword{internal}
