% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pb}
\alias{pb}
\title{Progress Bar}
\usage{
pb(total, format = "[:bar] :percent")
}
\arguments{
\item{total}{total number of ticks}

\item{format}{The format of the progress bar}
}
\value{
a function
}
\description{
Progress Bar
}
