% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_predict.R
\name{predict_classification}
\alias{predict_classification}
\title{Predict Classification}
\usage{
predict_classification(
  text,
  words,
  model,
  has_number = TRUE,
  has_symbol = c(has_equals = "=")
)
}
\arguments{
\item{text}{the text to predict}

\item{words}{a vector of words to find}

\item{model}{the model to fit}

\item{has_number}{whether to code the presence of numbers}

\item{has_symbol}{a named vector of symbols to detect}
}
\value{
a vector of predictions
}
\description{
Predict Classification
}
