% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{psychsci}
\alias{psychsci}
\title{Psychologial Science Open Access Paper Set}
\format{
A list of 250 paper objects
}
\source{
\url{https://journals.sagepub.com/home/pss}
}
\usage{
psychsci
}
\description{
250 open access papers from Psychological Science.
}
\keyword{datasets}
