% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf2grobid.R
\name{pytacheck}
\alias{pytacheck}
\title{Process a paper using the Pytacheck API}
\usage{
pytacheck(
  file_path,
  api_url = "https://api.metacheck.app/get-paper-metadata/",
  api_key = "66808e327dc79d135ba18e051673d906"
)
}
\arguments{
\item{api_url}{Base URL of the API}

\item{filename}{Path to the PDF file}
}
\value{
A paper object
}
\description{
Process a paper using the Pytacheck API
}
\keyword{internal}
