% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-helpers.R
\name{report_table}
\alias{report_table}
\title{Display a Table in a Report}
\usage{
report_table(table, colwidths = "auto", maxrows = 2, escape = FALSE)
}
\arguments{
\item{table}{the data frame to show in a table, or a vector for a list}

\item{colwidths}{set column widths as a vector of px (number > 1) or percent (numbers <= 1)}

\item{maxrows}{if the table has more rows than this, paginate}

\item{escape}{whether or not to escape the DT (necessary if using raw html)}
}
\value{
the datatable
}
\description{
A function to display tables in reports.
}
\examples{
report_table(iris)
}
