% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retractionwatch.R
\name{retractionwatch}
\alias{retractionwatch}
\alias{rw}
\title{RetractionWatch data}
\format{
A data frame with 44784+ rows and 2 columns:
\describe{
\item{doi}{Document ID}
\item{retractionwatch}{Nature of note(s)}
}
}
\source{
\url{https://api.labs.crossref.org/data/retractionwatch}
}
\usage{
retractionwatch()

rw()
}
\value{
a data frame
}
\description{
DOIs and nature of statements from the RetractionWatch database. Use \code{rw_date()} to find the date it was downloaded, and \code{rw_update()} to update it.
}
\examples{
retractionwatch()
}
