% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{scroll_table}
\alias{scroll_table}
\title{Make Scroll Table}
\usage{
scroll_table(table, height = 200)
}
\arguments{
\item{table}{the data frame to show in a table, or a vector for a list}

\item{height}{the height of the scroll window}
}
\value{
the markdown R chunk to create this table
}
\description{
A helper function for making module reports.
}
\examples{
scroll_table(LETTERS)
}
