% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_predict.R
\name{text_features}
\alias{text_features}
\title{Text features}
\usage{
text_features(text, words, has_number = TRUE, has_symbol = c(has_equals = "="))
}
\arguments{
\item{text}{a vector of the text strings to extract features from}

\item{words}{a vector of words to find}

\item{has_number}{whether to code the presence of numbers}

\item{has_symbol}{a named vector of symbols to detect}
}
\value{
a data frame of features for each text
}
\description{
Text features
}
